refactor!: Setup handler & entity command handler instead of events (#35)
Rewrite the driver setup flow handling to use a callback handler instead
of events with manual WebSocket interactions in the client.

Rewrite the entity command handler to use a callback handler instead of
events. One can install a global entity command handler, or individual
ones per entity or entity types.

The callback handlers are modeled the same way as in the Python
integration library. If an integration driver doesn't install a callback
handler, then the old events are emitted. This should simplify
transitioning to the new handlers.
Attention: the old events are planned to be removed by the end of 2024!

Other changes:
- Fix missing options property in available entities.
- Add a new setup-flow example and clean up / simplify the simulated light
example.
- Add ava testing framework and first utility tests for language strings.